KAFKA-17970: Moving some share purgatory classes from core to share module
