+++ share/src/main/java/org/apache/kafka/server/share/fetch/DelayedShareFetchKey.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package kafka.server.share;
+package org.apache.kafka.server.share.fetch;
 
 import org.apache.kafka.server.purgatory.DelayedOperationKey;
